+++ .dojorc
@@ -59,10 +59,8 @@
 			"src/slider",
 			"src/snackbar",
 			"src/speed-dial",
-			"src/stack/Stack",
-			"src/stack/HStack",
+			"src/stack",
 			"src/stack/Spacer",
-			"src/stack/VStack",
 			"src/wizard",
 			"src/switch",
 			"src/tab-container",
