+++ src/stack/README.md
@@ -1,6 +1,3 @@
 # Stacks
 
-`@dojo/widgets/stack/Stack` is a layout component that enables creating stacks of elements or widgets with minimal effort and apply spacing between them. The `Stack` widget can layout either vertically or horizontally and the `Spacer` widget can be used to take up space within the layout or when passed children let the children take up the specified space.
-
-#
-For convenience there are two widgets specifically for laying out in a one direction, `@dojo/widgets/stack/HStack` and `@dojo/widgets/stack/VStack`. A `Stack` has a number of properties for customizing the layout.
+`@dojo/widgets/stack/Stack` is a layout component that enables creating stacks of elements or widgets with minimal effort and apply spacing between them. The `Stack` widget can layout either vertically or horizontally and the `Spacer` widget can be used to take up space within the layout or when passed children let the children take up the specified space. The `Stack` has a number of properties for customizing the layout.
