+++ src/stack/index.tsx
@@ -1,27 +1,17 @@
 import { tsx, create, isWNode } from '@dojo/framework/core/vdom';
 import theme from '../middleware/theme';
-import * as fixedCss from './styles/hstack.m.css';
-import * as css from '../theme/default/hstack.m.css';
-import Spacer from './Spacer';
+
+import * as fixedCss from './styles/stack.m.css';
+import * as css from '../theme/default/stack.m.css';
+
 import {
 	DNode,
 	WNodeFactory,
 	OptionalWNodeFactory,
 	DefaultChildrenWNodeFactory
 } from '@dojo/framework/core/interfaces';
-
-export interface HStackProperties {
-	/** The vertical alignment of the stack */
-	align?: 'start' | 'middle' | 'end';
-	/** The spacing between children */
-	spacing?: 'small' | 'medium' | 'large';
-	/** The padding for stack container */
-	padding?: 'small' | 'medium' | 'large';
-	/** Stretches the container to fit the space */
-	stretch?: boolean;
-}
-
-const factory = create({ theme }).properties<HStackProperties>();
+import Spacer from './Spacer';
+export { Spacer } from './Spacer';
 
 function typeOf(
 	node: DNode,
@@ -31,8 +21,23 @@ function typeOf(
 	return isWNode(node) && compareTo.widgetConstructor === node.widgetConstructor;
 }
 
-export const HStack = factory(function HStack({ properties, middleware: { theme }, children }) {
-	const { align, spacing, padding = false, stretch = false } = properties();
+export interface StackProperties {
+	/** The direction of the stack */
+	direction: 'vertical' | 'horizontal';
+	/** The sets the alignment of the stack for the opposite direction of the stack, i.e. for a vertical stack it is the horizontal alignment */
+	align?: 'start' | 'middle' | 'end';
+	/** The spacing between children */
+	spacing?: 'small' | 'medium' | 'large';
+	/** The padding for stack container */
+	padding?: 'small' | 'medium' | 'large';
+	/** Stretches the container to fit the space */
+	stretch?: boolean;
+}
+
+const factory = create({ theme }).properties<StackProperties>();
+
+export const Stack = factory(function Stack({ properties, middleware: { theme }, children }) {
+	const { direction, align, spacing, padding = false, stretch = false } = properties();
 	const themeCss = theme.classes(css);
 	let spacingClass: string | undefined;
 	let paddingClass: string | undefined;
@@ -66,13 +71,13 @@ export const HStack = factory(function HStack({ properties, middleware: { theme 
 	}
 	switch (align) {
 		case 'start':
-			alignClass = fixedCss.top;
+			alignClass = fixedCss.start;
 			break;
 		case 'middle':
 			alignClass = fixedCss.middle;
 			break;
 		case 'end':
-			alignClass = fixedCss.bottom;
+			alignClass = fixedCss.end;
 			break;
 		default:
 			break;
@@ -82,15 +87,21 @@ export const HStack = factory(function HStack({ properties, middleware: { theme 
 		if (typeOf(child, Spacer)) {
 			return child;
 		}
-		return <div classes={[spacingClass, fixedCss.child]}>{child}</div>;
+		return (
+			<div classes={[spacingClass, direction === 'vertical' && alignClass, fixedCss.child]}>
+				{child}
+			</div>
+		);
 	});
 
 	return (
 		<div
 			classes={[
 				theme.variant(),
+				direction === 'horizontal' ? themeCss.horizontal : themeCss.vertical,
+				direction === 'horizontal' ? fixedCss.horizontal : fixedCss.vertical,
 				fixedCss.root,
-				alignClass,
+				direction === 'horizontal' && alignClass,
 				paddingClass,
 				stretch && fixedCss.stretch
 			]}
@@ -100,4 +111,4 @@ export const HStack = factory(function HStack({ properties, middleware: { theme 
 	);
 });
 
-export default HStack;
+export default Stack;
