--- src/stack/styles/hstack.m.css
@@ -1,30 +1,0 @@
-.root {
-	display: flex;
-	flex-direction: row;
-	justify-content: center;
-	height: 100%;
-}
-
-.stretch {
-	width: 100%;
-}
-
-.top {
-	align-items: flex-start;
-}
-
-.middle {
-	align-items: center;
-}
-
-.bottom {
-	align-items: flex-end;
-}
-
-.stretch > .child {
-	width: 100%;
-}
-
-.child:last-child {
-	margin-right: 0;
-}
