--- src/stack/styles/hstack.m.css.d.ts
@@ -1,6 +1,0 @@
-export const root: string;
-export const stretch: string;
-export const top: string;
-export const middle: string;
-export const bottom: string;
-export const child: string;
