+++ src/stack/styles/stack.m.css.d.ts
@@ -1,0 +1,8 @@
+export const root: string;
+export const vertical: string;
+export const horizontal: string;
+export const stretch: string;
+export const child: string;
+export const start: string;
+export const middle: string;
+export const end: string;
