--- src/stack/styles/vstack.m.css
@@ -1,37 +1,0 @@
-.root {
-	display: flex;
-	flex-direction: column;
-	width: 100%;
-}
-
-.stretch {
-	height: 100%;
-}
-
-.child.left {
-	align-items: flex-start;
-}
-
-.child.center {
-	align-items: center;
-}
-
-.child.right {
-	align-items: flex-end;
-}
-
-.child {
-	display: flex;
-	flex-direction: column;
-	width: 100%;
-	align-items: stretch;
-}
-
-.child:last-child {
-	margin-bottom: 0;
-	margin-bottom: auto;
-}
-
-.child:first-child {
-	margin-top: auto;
-}
