--- src/stack/styles/vstack.m.css.d.ts
@@ -1,6 +1,0 @@
-export const root: string;
-export const stretch: string;
-export const child: string;
-export const left: string;
-export const center: string;
-export const right: string;
