+++ src/stack/tests/HStack.spec.tsx
@@ -2,14 +2,14 @@ const { it, describe } = intern.getInterface('bdd');
 
 import { tsx } from '@dojo/framework/core/vdom';
 import { renderer, assertion } from '@dojo/framework/testing/renderer';
-import HStack from '../HStack';
-import * as fixedCss from '../styles/hstack.m.css';
-import * as css from '../../theme/default/hstack.m.css';
+import Stack from '../index';
+import * as fixedCss from '../styles/stack.m.css';
+import * as css from '../../theme/default/stack.m.css';
 import Spacer from '../Spacer';
 
 describe('Stacks - HStack Rendering', () => {
 	it('As Default', () => {
-		const r = renderer(() => <HStack>HStack Child</HStack>);
+		const r = renderer(() => <Stack direction="horizontal">HStack Child</Stack>);
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, undefined, false]}>
 				<div classes={[undefined, fixedCss.child]}>HStack Child</div>
@@ -19,7 +19,11 @@ describe('Stacks - HStack Rendering', () => {
 	});
 
 	it('With Small Spacing', () => {
-		const r = renderer(() => <HStack spacing="small">HStack Child</HStack>);
+		const r = renderer(() => (
+			<Stack direction="horizontal" spacing="small">
+				HStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, undefined, false]}>
 				<div classes={[css.smallSpacing, fixedCss.child]}>HStack Child</div>
@@ -29,7 +33,11 @@ describe('Stacks - HStack Rendering', () => {
 	});
 
 	it('With Medium Spacing', () => {
-		const r = renderer(() => <HStack spacing="medium">HStack Child</HStack>);
+		const r = renderer(() => (
+			<Stack direction="horizontal" spacing="medium">
+				HStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, undefined, false]}>
 				<div classes={[css.mediumSpacing, fixedCss.child]}>HStack Child</div>
@@ -39,7 +47,11 @@ describe('Stacks - HStack Rendering', () => {
 	});
 
 	it('With Large Spacing', () => {
-		const r = renderer(() => <HStack spacing="large">HStack Child</HStack>);
+		const r = renderer(() => (
+			<Stack direction="horizontal" spacing="large">
+				HStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, undefined, false]}>
 				<div classes={[css.largeSpacing, fixedCss.child]}>HStack Child</div>
@@ -49,7 +61,11 @@ describe('Stacks - HStack Rendering', () => {
 	});
 
 	it('With Small Padding', () => {
-		const r = renderer(() => <HStack padding="small">HStack Child</HStack>);
+		const r = renderer(() => (
+			<Stack direction="horizontal" padding="small">
+				HStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, css.smallPadding, false]}>
 				<div classes={[null, fixedCss.child]}>HStack Child</div>
@@ -59,7 +75,11 @@ describe('Stacks - HStack Rendering', () => {
 	});
 
 	it('With Medium Padding', () => {
-		const r = renderer(() => <HStack padding="medium">HStack Child</HStack>);
+		const r = renderer(() => (
+			<Stack direction="horizontal" padding="medium">
+				HStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, css.mediumPadding, false]}>
 				<div classes={[null, fixedCss.child]}>HStack Child</div>
@@ -69,7 +89,11 @@ describe('Stacks - HStack Rendering', () => {
 	});
 
 	it('With Large Padding', () => {
-		const r = renderer(() => <HStack padding="large">HStack Child</HStack>);
+		const r = renderer(() => (
+			<Stack direction="horizontal" padding="large">
+				HStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, css.largePadding, false]}>
 				<div classes={[null, fixedCss.child]}>HStack Child</div>
@@ -79,9 +103,13 @@ describe('Stacks - HStack Rendering', () => {
 	});
 
 	it('With Start Alignment', () => {
-		const r = renderer(() => <HStack align="start">HStack Child</HStack>);
+		const r = renderer(() => (
+			<Stack direction="horizontal" align="start">
+				HStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
-			<div classes={[null, fixedCss.root, fixedCss.top, undefined, false]}>
+			<div classes={[null, fixedCss.root, fixedCss.start, undefined, false]}>
 				<div classes={[null, fixedCss.child]}>HStack Child</div>
 			</div>
 		));
@@ -89,7 +117,11 @@ describe('Stacks - HStack Rendering', () => {
 	});
 
 	it('With Middle Alignment', () => {
-		const r = renderer(() => <HStack align="middle">HStack Child</HStack>);
+		const r = renderer(() => (
+			<Stack direction="horizontal" align="middle">
+				HStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, fixedCss.middle, undefined, false]}>
 				<div classes={[null, fixedCss.child]}>HStack Child</div>
@@ -99,9 +131,13 @@ describe('Stacks - HStack Rendering', () => {
 	});
 
 	it('With End Alignment', () => {
-		const r = renderer(() => <HStack align="end">HStack Child</HStack>);
+		const r = renderer(() => (
+			<Stack direction="horizontal" align="end">
+				HStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
-			<div classes={[null, fixedCss.root, fixedCss.bottom, undefined, false]}>
+			<div classes={[null, fixedCss.root, fixedCss.end, undefined, false]}>
 				<div classes={[null, fixedCss.child]}>HStack Child</div>
 			</div>
 		));
@@ -109,7 +145,11 @@ describe('Stacks - HStack Rendering', () => {
 	});
 
 	it('With Stretch', () => {
-		const r = renderer(() => <HStack stretch>HStack Child</HStack>);
+		const r = renderer(() => (
+			<Stack direction="horizontal" stretch>
+				HStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, undefined, fixedCss.stretch]}>
 				<div classes={[null, fixedCss.child]}>HStack Child</div>
@@ -120,11 +160,11 @@ describe('Stacks - HStack Rendering', () => {
 
 	it('With a Spacer', () => {
 		const r = renderer(() => (
-			<HStack padding="large">
+			<Stack direction="horizontal" padding="large">
 				<Spacer />
 				HStack Child
 				<Spacer />
-			</HStack>
+			</Stack>
 		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, css.largePadding, false]}>
