+++ src/stack/tests/VStack.spec.tsx
@@ -2,14 +2,14 @@ const { it, describe } = intern.getInterface('bdd');
 
 import { tsx } from '@dojo/framework/core/vdom';
 import { renderer, assertion } from '@dojo/framework/testing/renderer';
-import VStack from '../VStack';
-import * as fixedCss from '../styles/vstack.m.css';
-import * as css from '../../theme/default/vstack.m.css';
+import Stack from '../index';
+import * as fixedCss from '../styles/stack.m.css';
+import * as css from '../../theme/default/stack.m.css';
 import Spacer from '../Spacer';
 
 describe('Stacks - VStack Rendering', () => {
 	it('As Default', () => {
-		const r = renderer(() => <VStack>VStack Child</VStack>);
+		const r = renderer(() => <Stack direction="vertical">VStack Child</Stack>);
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, false]}>
 				<div classes={[null, null, fixedCss.child]}>VStack Child</div>
@@ -19,7 +19,11 @@ describe('Stacks - VStack Rendering', () => {
 	});
 
 	it('With Small Spacing', () => {
-		const r = renderer(() => <VStack spacing="small">VStack Child</VStack>);
+		const r = renderer(() => (
+			<Stack direction="vertical" spacing="small">
+				VStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, false]}>
 				<div classes={[css.smallSpacing, null, fixedCss.child]}>VStack Child</div>
@@ -29,7 +33,11 @@ describe('Stacks - VStack Rendering', () => {
 	});
 
 	it('With Medium Spacing', () => {
-		const r = renderer(() => <VStack spacing="medium">VStack Child</VStack>);
+		const r = renderer(() => (
+			<Stack direction="vertical" spacing="medium">
+				VStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, false]}>
 				<div classes={[css.mediumSpacing, null, fixedCss.child]}>VStack Child</div>
@@ -39,7 +47,11 @@ describe('Stacks - VStack Rendering', () => {
 	});
 
 	it('With Large Spacing', () => {
-		const r = renderer(() => <VStack spacing="large">VStack Child</VStack>);
+		const r = renderer(() => (
+			<Stack direction="vertical" spacing="large">
+				VStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, false]}>
 				<div classes={[css.largeSpacing, null, fixedCss.child]}>VStack Child</div>
@@ -49,7 +61,11 @@ describe('Stacks - VStack Rendering', () => {
 	});
 
 	it('With Small Padding', () => {
-		const r = renderer(() => <VStack padding="small">VStack Child</VStack>);
+		const r = renderer(() => (
+			<Stack direction="vertical" padding="small">
+				VStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, css.smallPadding, false]}>
 				<div classes={[null, null, fixedCss.child]}>VStack Child</div>
@@ -59,7 +75,11 @@ describe('Stacks - VStack Rendering', () => {
 	});
 
 	it('With Medium Padding', () => {
-		const r = renderer(() => <VStack padding="medium">VStack Child</VStack>);
+		const r = renderer(() => (
+			<Stack direction="vertical" padding="medium">
+				VStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, css.mediumPadding, false]}>
 				<div classes={[null, null, fixedCss.child]}>VStack Child</div>
@@ -69,7 +89,11 @@ describe('Stacks - VStack Rendering', () => {
 	});
 
 	it('With Large Padding', () => {
-		const r = renderer(() => <VStack padding="large">VStack Child</VStack>);
+		const r = renderer(() => (
+			<Stack direction="vertical" padding="large">
+				VStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, css.largePadding, false]}>
 				<div classes={[null, null, fixedCss.child]}>VStack Child</div>
@@ -79,37 +103,53 @@ describe('Stacks - VStack Rendering', () => {
 	});
 
 	it('With Start Alignment', () => {
-		const r = renderer(() => <VStack align="start">VStack Child</VStack>);
+		const r = renderer(() => (
+			<Stack direction="vertical" align="start">
+				VStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, false]}>
-				<div classes={[null, fixedCss.left, fixedCss.child]}>VStack Child</div>
+				<div classes={[null, fixedCss.start, fixedCss.child]}>VStack Child</div>
 			</div>
 		));
 		r.expect(baseAssertion);
 	});
 
 	it('With Middle Alignment', () => {
-		const r = renderer(() => <VStack align="middle">VStack Child</VStack>);
+		const r = renderer(() => (
+			<Stack direction="vertical" align="middle">
+				VStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, false]}>
-				<div classes={[null, fixedCss.center, fixedCss.child]}>VStack Child</div>
+				<div classes={[null, fixedCss.middle, fixedCss.child]}>VStack Child</div>
 			</div>
 		));
 		r.expect(baseAssertion);
 	});
 
 	it('With End Alignment', () => {
-		const r = renderer(() => <VStack align="end">VStack Child</VStack>);
+		const r = renderer(() => (
+			<Stack direction="vertical" align="end">
+				VStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, false]}>
-				<div classes={[null, fixedCss.right, fixedCss.child]}>VStack Child</div>
+				<div classes={[null, fixedCss.end, fixedCss.child]}>VStack Child</div>
 			</div>
 		));
 		r.expect(baseAssertion);
 	});
 
 	it('With Stretch', () => {
-		const r = renderer(() => <VStack stretch>VStack Child</VStack>);
+		const r = renderer(() => (
+			<Stack direction="vertical" stretch>
+				VStack Child
+			</Stack>
+		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, undefined, fixedCss.stretch]}>
 				<div classes={[null, null, fixedCss.child]}>VStack Child</div>
@@ -120,11 +160,11 @@ describe('Stacks - VStack Rendering', () => {
 
 	it('With a Spacer', () => {
 		const r = renderer(() => (
-			<VStack padding="large">
+			<Stack direction="vertical" padding="large">
 				<Spacer />
 				VStack Child
 				<Spacer />
-			</VStack>
+			</Stack>
 		));
 		const baseAssertion = assertion(() => (
 			<div classes={[null, fixedCss.root, css.largePadding, false]}>
