+++ src/theme/default/index.ts
@@ -22,7 +22,7 @@ import * as gridPlaceholderRow from './grid-placeholder-row.m.css';
 import * as gridRow from './grid-row.m.css';
 import * as grid from './grid.m.css';
 import * as helperText from './helper-text.m.css';
-import * as hstack from './hstack.m.css';
+import * as stack from './stack.m.css';
 import * as icon from './icon.m.css';
 import * as label from './label.m.css';
 import * as listItem from './list-item.m.css';
@@ -58,7 +58,6 @@ import * as tree from './tree.m.css';
 import * as tooltip from './tooltip.m.css';
 import * as twoColumnLayout from './two-column-layout.m.css';
 import * as typeahead from './typeahead.m.css';
-import * as vstack from './vstack.m.css';
 import * as defaultVariant from './variants/default.m.css';
 
 export default {
@@ -122,8 +121,7 @@ export default {
 		'@dojo/widgets/tooltip': tooltip,
 		'@dojo/widgets/two-column-layout': twoColumnLayout,
 		'@dojo/widgets/typeahead': typeahead,
-		'@dojo/widgets/hstack': hstack,
-		'@dojo/widgets/vstack': vstack
+		'@dojo/widgets/stack': stack
 	},
 	variants: {
 		default: defaultVariant
