--- src/theme/default/vstack.m.css
@@ -1,29 +1,0 @@
-/* Defines the small spacing size */
-.smallSpacing {
-	margin-bottom: 8px;
-}
-
-/* Defines the medium spacing size */
-.mediumSpacing {
-	margin-bottom: 12px;
-}
-
-/* Defines the large spacing size */
-.largeSpacing {
-	margin-bottom: 16px;
-}
-
-/* Defines the small padding size */
-.smallPadding {
-	padding: 8px;
-}
-
-/* Defines the medium padding size */
-.mediumPadding {
-	padding: 12px;
-}
-
-/* Defines the large padding size */
-.largePadding {
-	padding: 16px;
-}
