--- src/theme/dojo/hstack.m.css
@@ -1,23 +1,0 @@
-.smallSpacing {
-	margin-right: var(--mdc-hstack-spacing-small);
-}
-
-.mediumSpacing {
-	margin-right: var(--mdc-hstack-spacing-medium);
-}
-
-.largeSpacing {
-	margin-right: var(--mdc-hstack-spacing-large);
-}
-
-.smallPadding {
-	padding: var(--mdc-hstack-padding-small);
-}
-
-.mediumPadding {
-	padding: var(--mdc-hstack-padding-medium);
-}
-
-.largePadding {
-	padding: var(--mdc-hstack-padding-large);
-}
