+++ src/theme/dojo/stack.m.css
@@ -1,0 +1,67 @@
+/* Class set when in vertical mode */
+.vertical {
+}
+
+/* Class set when in horizontal mode */
+.horizontal {
+}
+
+/* Defines the small spacing size */
+.horizontal > .smallSpacing {
+	margin-right: var(--vstack-spacing-small);
+}
+
+/* Defines the medium spacing size */
+.horizontal > .mediumSpacing {
+	margin-right: var(--vstack-spacing-medium);
+}
+
+/* Defines the large spacing size */
+.horizontal > .largeSpacing {
+	margin-right: var(--vstack-spacing-large);
+}
+
+/* Defines the small padding size */
+.horizontal.smallPadding {
+	padding: var(--hstack-padding-small);
+}
+
+/* Defines the medium padding size */
+.horizontal.mediumPadding {
+	padding: var(--hstack-padding-medium);
+}
+
+/* Defines the large padding size */
+.horizontal.largePadding {
+	padding: var(--hstack-padding-large);
+}
+
+/* Defines the small spacing size */
+.vertical > .smallSpacing {
+	margin-bottom: var(--vstack-spacing-small);
+}
+
+/* Defines the medium spacing size */
+.vertical > .mediumSpacing {
+	margin-bottom: var(--vstack-spacing-medium);
+}
+
+/* Defines the large spacing size */
+.vertical > .largeSpacing {
+	margin-bottom: var(--vstack-spacing-large);
+}
+
+/* Defines the small padding size */
+.vertical.smallPadding {
+	padding: var(--vstack-padding-small);
+}
+
+/* Defines the medium padding size */
+.vertical.mediumPadding {
+	padding: var(--vstack-padding-medium);
+}
+
+/* Defines the large padding size */
+.vertical.largePadding {
+	padding: var(--vstack-padding-large);
+}
