--- src/theme/dojo/vstack.m.css
@@ -1,23 +1,0 @@
-.smallSpacing {
-	margin-bottom: var(--vstack-spacing-small);
-}
-
-.mediumSpacing {
-	margin-bottom: var(--vstack-spacing-medium);
-}
-
-.largeSpacing {
-	margin-bottom: var(--vstack-spacing-large);
-}
-
-.smallPadding {
-	padding: var(--vstack-padding-small);
-}
-
-.mediumPadding {
-	padding: var(--vstack-padding-medium);
-}
-
-.largePadding {
-	padding: var(--vstack-padding-large);
-}
