+++ src/theme/material/index.ts
@@ -23,7 +23,6 @@ import * as grid from './grid.m.css';
 import * as headerCard from './header-card.m.css';
 import * as header from './header.m.css';
 import * as helperText from './helper-text.m.css';
-import * as hstack from './hstack.m.css';
 import * as icon from './icon.m.css';
 import * as label from './label.m.css';
 import * as listItem from './list-item.m.css';
@@ -60,7 +59,7 @@ import * as tooltip from './tooltip.m.css';
 import * as tree from './tree.m.css';
 import * as twoColumnLayout from './two-column-layout.m.css';
 import * as typeahead from './typeahead.m.css';
-import * as vstack from './vstack.m.css';
+import * as stack from './stack.m.css';
 import * as defaultVariant from './variants/default.m.css';
 import * as darkVariant from './variants/dark.m.css';
 
@@ -127,8 +126,7 @@ export default {
 		'@dojo/widgets/tree': tree,
 		'@dojo/widgets/two-column-layout': twoColumnLayout,
 		'@dojo/widgets/typeahead': typeahead,
-		'@dojo/widgets/hstack': hstack,
-		'@dojo/widgets/vstack': vstack
+		'@dojo/widgets/stack': stack
 	},
 	variants: {
 		default: defaultVariant,
