--- src/theme/material/vstack.m.css
@@ -1,23 +1,0 @@
-.smallSpacing {
-	margin-bottom: var(--mdc-vstack-spacing-small);
-}
-
-.mediumSpacing {
-	margin-bottom: var(--mdc-vstack-spacing-medium);
-}
-
-.largeSpacing {
-	margin-bottom: var(--mdc-vstack-spacing-large);
-}
-
-.smallPadding {
-	padding: var(--mdc-vstack-padding-small);
-}
-
-.mediumPadding {
-	padding: var(--mdc-vstack-padding-medium);
-}
-
-.largePadding {
-	padding: var(--mdc-vstack-padding-large);
-}
