--- src/theme/material/vstack.m.css.d.ts
@@ -1,6 +1,0 @@
-export const smallSpacing: string;
-export const mediumSpacing: string;
-export const largeSpacing: string;
-export const smallPadding: string;
-export const mediumPadding: string;
-export const largePadding: string;
